% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countBT}
\alias{countBT}
\title{RNAseq count data downsampled from GSE60424}
\format{A list with one count mixture (count) and a categorical vector
giving subtypes (group).}
\usage{
data(countBT)
}
\description{
Three cell subtypes (B cells, CD4+ T cells, CD8+ T cells) were purified from
20 fresh blood samples. RNA was extracted from each of these cell subsets
and processed into RNA sequencing libraries (Illumina TruSeq). Sequencing
libraries were analyzed on an Illumina HiScan, with a target read depth
of ~20M reads. Reads were demultiplexed, mapped to human gene
models (ENSEMBL), and tabulated using HTSeq.
We downsample the original data to 10000 genes.
Subtype labels for purified populations are also included.
(Data generation script can be found in ./data_raw folder.)
}
\references{
Linsley et al. PLoS One 2014;9(10):e109760. PMID: 25314013
}
