% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OVESEG.R
\docType{package}
\name{OVESEG-package}
\alias{OVESEG-package}
\alias{OVESEG}
\title{CAMTHC: A package for Tissue Heterogeneity Characterization.}
\description{
The core function in this package by combining molecular markers
from CAM and from prior knowledge to analyze mixture expressions.
}
