% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pairwise_tstat_unscaled}
\alias{pairwise_tstat_unscaled}
\title{Multiply a number by two}
\usage{
pairwise_tstat_unscaled(ymean, stdevUnscaled)
}
\arguments{
\item{y}{A single integer.}

\item{group}{A single integer.}
}
\description{
Multiply a number by two
}
