% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OVEtstatPermTopM.R
\name{permfunc}
\alias{permfunc}
\title{Internal function for one permutation task}
\usage{
permfunc(p, y, group, weights, alpha, combM, geneSubset, seeds)
}
\arguments{
\item{p}{an integer indicating permutation index}

\item{y}{an expressions matrix}

\item{group}{a integer vector indicating group labels}

\item{weights}{optional numeric matrix containing prior weights}

\item{alpha}{parameter specifying within-group variance estimator to be used}

\item{combM}{a integer matrix with each row giving one choice of M groups}

\item{geneSubset}{a integer vector indicating the probe pattern of combM}

\item{seed}{an integer seed for the random number generator}
}
\value{
test statistics and upregulated group indexes after one permutaion
}
\description{
Internal function for one permutation task
}
