% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pvalue_weighted_est}
\alias{pvalue_weighted_est}
\title{p-value by weighted estimation}
\usage{
pvalue_weighted_est(tt, ttperm, w)
}
\arguments{
\item{tt}{a vector of test statistics.}

\item{ttperm}{a matrix of test statistics from permutaitons.}

\item{W}{a matrix containing weights for each spot in ttperm.}
}
\description{
p-value by weighted estimation
}
