% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalueWeightedEst.R
\name{pvalueWeightedEst}
\alias{pvalueWeightedEst}
\title{pvalueWeightedEst}
\usage{
pvalueWeightedEst(tt, ttstar, w)
}
\arguments{
\item{tt}{A data set that will be internally coerced into a matrix.
Each row is a gene and each column is a sample. Missing values
are not supported.
All-zero rows will be removed internally.}

\item{ttstar}{The matrix whose rows are loading vectors}

\item{W}{The matrix whose rows are loading vectors}
}
\value{
The data after perspective projection.
}
\description{
This function reduces data dimension by loading matrix and then project
dimension-reduced data to the hyperplane orthogonal to c(1,0,...,0),
i.e., the first axis in the new coordinate system..
}
\details{
This function can project gene expression vectors to simplex
filtering. This function helps observe all genes in simplex plot.
}
\examples{
#obtain data
data(RocheBT)
data <- RocheBT$y

#preprocess data
}
