% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RocheBT}
\alias{RocheBT}
\title{mRNA expression data downsampled from GSE28490}
\format{A list with three matrices: mixture profiles (X), mixing proportions
(A) and pure profiles (S).}
\usage{
data(RocheBT)
}
\description{
Rat brain, liver and lung biospecimens derived from one animal at the cRNA
homogenate level in different proportions. 3 technical replicates each.
We downsample the original data to 10000 probes/probesets and 7 mixtures.
Proportions used in experiments and pure expression profiles are also
included.
}
\references{
Shen-Orr et al. (2010) Nat Methods 2010 Apr;7(4):287-9.
PMID: 20208531
}
