% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{row_min}
\alias{row_min}
\title{min value for each row}
\usage{
row_min(Y)
}
\arguments{
\item{Y}{a numeric matrix}
}
\value{
a numeric vector indicating min value in each row
}
\description{
min value for each row
}
\keyword{internal}
