% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{row_which_max}
\alias{row_which_max}
\title{Multiply a number by two}
\usage{
row_which_max(Y)
}
\arguments{
\item{y}{A single integer.}

\item{group}{A single integer.}
}
\description{
Multiply a number by two
}
