% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{shuffle_topm}
\alias{shuffle_topm}
\title{Multiply a number by two}
\usage{
shuffle_topm(y, group, weights, combM, geneSubset, seed)
}
\arguments{
\item{y}{A single integer.}

\item{group}{A single integer.}
}
\description{
Multiply a number by two
}
