% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pairwise_tstat_unscaled}
\alias{pairwise_tstat_unscaled}
\title{pairwise t-statistics (unscaled)}
\usage{
pairwise_tstat_unscaled(ymean, stdevUnscaled)
}
\arguments{
\item{ymean}{a numeric matrix containing group means.}

\item{stdevUnscaled}{a numeric matrix containing unscaled standard
deviations of the group means.}
}
\description{
pairwise t-statistics (unscaled)
}
\keyword{internal}
