% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RocheBT}
\alias{RocheBT}
\title{mRNA expression data downsampled from GSE28490 (Roche)}
\format{A list with one expression matrix (y) and a categorical vector
giving subtypes (group).}
\usage{
data(RocheBT)
}
\description{
Three cell subtypes (B cells, CD4+ T cells, CD8+ T cells) were isolated from
5 pools of 5 healthy donors each. RNA obtained from these 15 purified
populations were subsequently used for mRNA expression profiling by
HG-U133Plus2.0 microarrays. We downsample the original data to 5000
probes/probesets.
Subtype labels for purified populations are also included.
(Data generation script can be found in ./data_raw folder.)
}
\references{
Allantaz et al. PLoS One 2012;7(1):e29979. PMID: 22276136
}
