% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{row_which_max}
\alias{row_which_max}
\title{which.max for each row}
\usage{
row_which_max(Y)
}
\arguments{
\item{Y}{a numeric matrix}
}
\value{
a integer vector indicating the location of max value in each row
}
\description{
which.max for each row
}
\keyword{internal}
