% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{shuffle_topm}
\alias{shuffle_topm}
\title{Shuffle the top M groups}
\usage{
shuffle_topm(y, group, weights, combM, geneSubset, seed)
}
\arguments{
\item{y}{a numeric matrix to be shuffled.}

\item{group}{a integer vector indicating group indexes.}

\item{weights}{optional numeric matrix containing prior weights.}

\item{combM}{a integer matrix with each row giving one choice of M groups}

\item{geneSubset}{a integer vector indicating the probe pattern of combM}

\item{seed}{an integer seed for the random number generator.}
}
\description{
Shuffle the top M groups
}
\keyword{internal}
